m4naos: drivers: use printf() format specifiers
This is for testing the libc integration and our init levels at the
same time. Ideally, we would use these different init levels to
initialize drivers in order without having to explicitly export a
symbol and call it in the right place.

The goal is to have early init code traverse the sections in order and
for each section call each of the function pointers placed there by
the linker.

Signed-off-by: Felipe Balbi <felipe@balbi.sh>