m4naos: improve the way sources get added to target
Instead of manually appending to a global SRCS variable, we can make
use of target_sources() to append more sources to the executable. This
is standard CMake feature.

We increase minimum required cmake to make sure we have policy CMP0076
enable. That was enabled by default in CMake 3.13 and above, but to be
safe we require a minimum 3.16 seen that the minimal version on my
system is 3.19 at the moment of this writing.

Signed-off-by: Felipe Balbi <felipe@balbi.sh>