m4naos: rcc: test memory allocation
Now that newlib nano is integrated, we can start allocating memory on
the heap. With this, we can start to introduce a more dynamic approach
to driver init.

The proposed sequence, for now, is:

- Initialize sysclk (before kernel_init)
- Initialize clock driver (core)
- Initialize UART driver (postcore)

More drivers will come in the future, but this is a good starting
point. Note that this commit merely tests memory allocation by
allocating a small context structure for RCC (clock) driver.

Signed-off-by: Felipe Balbi <felipe@balbi.sh>