m4naos: start testing drivers and devices
We're defining a const table for all devices available in the core and
maintaining a list of drivers. During initialization we will traverse
the array and list to match them by name.

If the driver name is a substring of the device name, we consider them
to match. This allows for a situation where the device name is
e.g. 'timer2' but the generic driver is named 'timer'.

Signed-off-by: Felipe Balbi <felipe@balbi.sh>