m4naos: place table of devices in CCM
We want the devices themselves to be modifiable, but we don't want to
waste the little SRAM we have. A good alternative is to use the CCM
for the device table.

In general, this will be a good place to put any data describing the
HW, for example clock and reset description for the devices could go
in CCM as well.

Signed-off-by: Felipe Balbi <felipe@balbi.sh>