gpio: pass IRQ number as a resource
While at that, also make sure to toggle a pin on the same GPIO
controller as we don't yet have means to control GPIOs from different
blocks.

A downside of this commit is that to "see" the IRQ working, we must
use a scope or connect an LED to PA1 (EXT2-8 on Olimex H405).

Signed-off-by: Felipe Balbi <felipe@balbi.sh>