main: make tasks priviledged
Because of local_irq_enable(), any task that needs to print to console
must be privileged. This will be fixed in the future by adding an SVC
call for console printing. Of course, this is assuming that we will
use UART in the long run, but the goal is to use USB for any and every
communication.

Signed-off-by: Felipe Balbi <felipe@balbi.sh>